Add switch tab event listeners
Here I grabbed reference to each nav menu item, then added a switchTab function as a click event listener function to each. switchTab takes a tabRender function (on of the tab render I have as the default export of each tab module). switchTab first changes the background of each menu item to mistyrose (the body background) before running the render. This is because each render sets its respective menu item background to white so that color persists on each tab, making it obvious which tab the user is one. By setting each item to mistyRose with switchTab, that white color change is not permanent after a tab is switched to the first time.

At this point, the project is "done". I say "done" because while I've met all of the project requirements, the final product is pretty lackluster. It just meets the requirements but does not really go above or beyond anywhere. The content and design leaves much to be desired. If I were to use this for a portfolio project (I doubt I would), I would then make it much better, but at this point I'm going to call it a day and move on with the class.